Fix bug where conditions caused dkms to be removed and in turn zfs
The issue was trigged by adding virtualbox-dkms to the absent list. A
condition exists that will include the package dkms whenever
virtualbox-dkms is included. This logic also worked for removing
packages. Also using the condition logic when removing is probably a bad
idea and I thus stop doing it hereby.